Fix 2.6.19 Loft eth1 oops on 0.3.1 firmware due to incorrect rxdoneq_id setup. THX to Christian for helping to identify the issue. THX to Gordon for the initial backport. The testing loft_defconfig is included.
Eth0 worked, but that was more luck than proper initialization. 
The rxdoneq_id had a value of 0 instead of 4 as in ixdp425-setup.c. 
The value for the other targets were corrected to match.  
The Loft oops was corrected and both eth0 and eth1 properly 
initialize now. The other target values were corrected, but not
tested.